Re-organize under org.specs2.cats package
